backend/requirements.txt
fastapi
uvicorn
transformers
faiss-cpu
boto3
torch
datasets